merges from trunk
git-svn-id: file:///Users/petejw/Documents/libmesh_svn_bak@5604 434f946d-2f3d-0410-ba4c-cb9f52fb0dbf

+++ branches/libmesh.automake/contrib/bin/buildall.sh
@@ -1,22 +1,28 @@
 #!/bin/bash
 
-./bootstrap
+#./bootstrap
+
+if (test "x$AEROLAB_SYSTEM_CLASS" = "x"); then
+    echo "ERROR:  expecting a AEROLAB_SYSTEM_CLASS env var!"
+    exit 1
+fi
 
 top_dir=`pwd`
-install_dir=/lustre/work/benkirk/codes/install/aerolab_workstations
+install_dir=/lustre/work/benkirk/codes/install/$AEROLAB_SYSTEM_CLASS
 
 rm -rf $install_dir
 
-for METHOD in dbg devel opt ; do
+for METHOD in opt devel dbg ; do
 
     cd $top_dir
-    rm -rf $METHOD && mkdir $METHOD && cd $METHOD
+    builddir=$METHOD-$AEROLAB_SYSTEM_CLASS
+    rm -rf $builddir && mkdir $builddir && cd $builddir
 
     echo " "
-    echo " Configuring via ../configure --prefix=$install_dir $@"
+    echo " Configuring METHOD=$METHOD via ../configure --with-cxx=`which mpicxx` --with-cc=`which mpicc` --with-f77=`which mpif77` --with-fc=`which mpif90` --prefix=$install_dir $@"
     echo " "
 
-    ../configure --prefix=$install_dir $@
-    make -j 24 install
+    ../configure --with-cxx=`which mpicxx` --with-cc=`which mpicc` --with-f77=`which mpif77` --with-fc=`which mpif90` --prefix=$install_dir $@
+    make -j 8 install
 done
